OAK-4838: Move S3 classes to oak-blob-cloud module
Refactor packages for common classes

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1762416 13f79535-47bb-0310-9956-ffa450edef68
